Changed: Move the existing bc2 transform into transforms/standard, to match BC1
