+++ projects/dxt-lossless-transform-bc2/benches/unsplit_blocks/avx512.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc2::split_blocks::unsplit::avx512::avx512_shuffle;
+use dxt_lossless_transform_bc2::transforms::standard::untransform::avx512::avx512_shuffle;
 use safe_allocator_api::RawAlloc;
 
 fn bench_shuffle(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
