+++ projects/dxt-lossless-transform-bc2/benches/unsplit_blocks/portable32.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc2::split_blocks::unsplit::*;
+use dxt_lossless_transform_bc2::transforms::standard::untransform::*;
 use safe_allocator_api::RawAlloc;
 
 fn bench_portable32(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
