+++ projects/dxt-lossless-transform-bc2/src/transforms/standard/mod.rs
@@ -60,8 +60,8 @@
 //! This is possible to do with either allocating half of a buffer, and then copying the other half back,
 //! or outputting it all to a single buffer. Outputting all to single buffer is faster.
 //!
-pub mod split;
-pub mod unsplit;
+pub mod transform;
+pub mod untransform;
 
 /// Transform BC2 data from standard interleaved format to separated alpha/color/index format
 /// using the best known implementation for the current CPU.
@@ -75,7 +75,7 @@ pub mod unsplit;
 #[inline]
 pub unsafe fn split_blocks(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
     debug_assert!(len % 16 == 0);
-    split::split_blocks(input_ptr, output_ptr, len);
+    transform::split_blocks(input_ptr, output_ptr, len);
 }
 
 /// Transform BC2 data from separated alpha/color/index format back to standard interleaved format
@@ -90,5 +90,5 @@ pub unsafe fn split_blocks(input_ptr: *const u8, output_ptr: *mut u8, len: usize
 #[inline]
 pub unsafe fn unsplit_blocks(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
     debug_assert!(len % 16 == 0);
-    unsplit::unsplit_blocks(input_ptr, output_ptr, len);
+    untransform::unsplit_blocks(input_ptr, output_ptr, len);
 }
