+++ projects/dxt-lossless-transform-bc2/src/transforms/standard/transform/portable32.rs
